Removed W and S
So I just recently noticed that if you press the W or S key, the player's animation for moving will still run. So instead of writing code for that, I just deleted the W and S key from the player input action map. I can add them back if we need them however.