Burststep Update 1
Changed the burst step so the player's sprite flips whenever the player tries to burst step making it so their back is facing the direction they're burst stepping in making it more like a backstep. Also added a camera shake singleton which is called when the player burst steps for now. Also added a trail renderer to show when player is either dashing or burst stepping.